pyproject.toml
[tool.poetry]
name = "tiling"
version = "0.1.0"
description = "Create tile manifest for pathology slides"
authors = ["Raymond Lim <limr@mskcc.org>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ConfigArgParse = "^1.7"
fsspec = "^2023.12.2"
pandas = "^2.1.4"
numpy = "^1.26.3"
loguru = "^0.7.2"
pandera = "^0.18.0"
tiffslide = "^2.3.1"
pillow = "^10.2.0"
scikit-image = "^0.22.0"
dask = {extras = ["distributed"], version = "^2024.1.0"}
staintools = "^2.1.2"
spams = "^2.6.5.4"
pyarrow = "^15.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
mkdocs = "^1.5.3"
isort = "^5.13.2"
black = "^23.12.1"
flake8 = "^7.0.0"
pydocstyle = "^6.3.0"
ipython = "^8.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
create_tile_manifest = 'tiling.cli:main'
